Ajustes Rol Admin

+++ zoo/src/main/java/com/nelumbo/zoo/services/impl/UserServiceImpl.java
@@ -4,11 +4,13 @@ import com.nelumbo.zoo.dtos.UserDTO;
 import com.nelumbo.zoo.entities.UserEntity;
 import com.nelumbo.zoo.repositories.UserRepository;
 import com.nelumbo.zoo.services.UserService;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -18,22 +20,30 @@ public class UserServiceImpl implements UserService {
 
     public List<UserDTO> getUsers() {
         List<UserEntity> userEntities = (List<UserEntity>) userRepository.findAll();
-        List<UserDTO> userDTOs = userEntities.stream()
+        return userEntities.stream()
                 .map(userEntity ->
                         UserDTO.builder()
                                 .id(userEntity.getId())
                                 .name(userEntity.getName())
                                 .email(userEntity.getEmail())
-                                .pass(userEntity.getPass())
                                 .role(userEntity.getRole())
                                 .build()
                 )
                 .collect(Collectors.toList());
-        return userDTOs;
     }
 
+    @Transactional
+    public UserDTO saveUser(UserDTO userDTO) {
+        if (isNullOrEmpty(userDTO.getName()) || isNullOrEmpty(userDTO.getEmail()) || isNullOrEmpty(userDTO.getPass())
+         || isNullOrEmpty(userDTO.getRole())) {
+            throw new IllegalArgumentException("Los campos 'name', 'email', 'pass' y 'role' son requeridos.");
+        }
 
-    public UserDTO saveUser (UserDTO userDTO) {
+        Optional<UserEntity> existingUser = userRepository.findOneByEmail(userDTO.getEmail());
+        if (existingUser.isPresent()) {
+            throw new IllegalArgumentException("El correo electrónico ya está en uso.");
+        }
+
         UserEntity userEntity = UserEntity
                 .builder()
                 .id(userDTO.getId())
@@ -47,7 +57,18 @@ public class UserServiceImpl implements UserService {
         return userDTO;
     }
 
-    public boolean deleteUser (Long id) {
+    public boolean deleteUser(Long id) {
+        UserEntity userToDelete = userRepository.findById(id).orElse(null);
+        if (userToDelete == null) {
+            return false;
+        }
+        if ("ADMIN".equals(userToDelete.getRole())) {
+            long adminCount = userRepository.countByRole("ADMIN");
+            if (adminCount <= 1) {
+                throw new IllegalArgumentException("No se ha podido eliminar el usuario debido a que es el unico" +
+                        " administrador registrado.");
+            }
+        }
         try {
             userRepository.deleteById(id);
             return true;
@@ -77,5 +98,9 @@ public class UserServiceImpl implements UserService {
 
         userRepository.save(user1);
         userRepository.save(user2);
+    }
+
+    private boolean isNullOrEmpty(String str) {
+        return str == null || str.trim().isEmpty();
     }
 }
